+++ thoth/messaging/adviser_trigger.py
@@ -19,6 +19,7 @@
 """This is Thoth Messaging module for AdviseTriggerMessage."""
 
 import logging
+import attr
 from typing import Any
 from typing import Optional
 from typing import Dict
@@ -33,27 +34,28 @@ class AdviserTriggerMessage(MessageBase):
 
     base_name = "thoth.adviser-trigger"
 
+    @attr.s
     class MessageContents(BaseMessageContents):  # type: ignore
         """Class used to represent contents of a message Kafka topic."""
 
-        application_stack: Dict[Any, Any]
-        recommendation_type: str
-        dev: bool = False
-        debug: bool = False
-        count: Optional[int] = None
-        limit: Optional[int] = None
-        runtime_environment: Optional[Dict[Any, Any]] = None
-        library_usage: Optional[Dict[Any, Any]] = None
-        origin: Optional[str] = None
-        job_id: Optional[str] = None
-        limit_latest_versions: Optional[int] = None
-        github_event_type: Optional[str] = None
-        github_check_run_id: Optional[int] = None
-        github_installation_id: Optional[int] = None
-        github_base_repo_url: Optional[str] = None
-        re_run_adviser_id: Optional[str] = None
-        source_type: Optional[str] = None
-        version: str = "v1"
+        application_stack = attr.ib(type=Dict[Any, Any])
+        recommendation_type = attr.ib(type=str)
+        dev = attr.ib(type=bool, default=False)
+        debug = attr.ib(type=bool, default=False)
+        count = attr.ib(type=Optional[int], default=None)
+        limit = attr.ib(type=Optional[int], default=None)
+        runtime_environment = attr.ib(type=Optional[Dict[Any, Any]], default=None)
+        library_usage = attr.ib(type=Optional[Dict[Any, Any]], default=None)
+        origin = attr.ib(type=Optional[str], default=None)
+        job_id = attr.ib(type=Optional[str], default=None)
+        limit_latest_versions = attr.ib(type=Optional[int], default=None)
+        github_event_type = attr.ib(type=Optional[str], default=None)
+        github_check_run_id = attr.ib(type=Optional[int], default=None)
+        github_installation_id = attr.ib(type=Optional[int], default=None)
+        github_base_repo_url = attr.ib(type=Optional[str], default=None)
+        re_run_adviser_id = attr.ib(type=Optional[str], default=None)
+        source_type = attr.ib(type=Optional[str], default=None)
+        version = attr.ib(type=str, default="v1", init=False)
 
     def __init__(self,):
         """Initialize advise-justification topic."""